+++ src/test/java/cart/controller/unit/ViewControllerUnitTest.java
@@ -1,5 +1,6 @@
-package cart.controller;
+package cart.controller.unit;
 
+import cart.controller.ViewController;
 import cart.dao.entity.MemberEntity;
 import cart.dao.entity.ProductEntity;
 import cart.dto.ResponseMemberDto;
